requirements.txt
# Playwriter - AI-driven narrative generation
# Run: pip install -r requirements.txt

# --- API framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# --- Data validation & config ---
pydantic>=2.0
pydantic-settings>=2.0

# --- LLM provider SDKs ---
openai>=1.10.0
anthropic>=0.40.0
groq>=0.5.0

# --- Database ---
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0

# --- Data processing ---
pandas>=2.0

# --- Environment ---
python-dotenv>=1.0.0
